requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.21

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# LINE Bot SDK
line-bot-sdk==3.6.0

# Scheduler
apscheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
